refactor(examples): Make the examples scenario / feature based
- Removed the placeholder main.go from the server example.
- Added a new streaming example demonstrating real-time AI responses.
- Implemented a streaming client that handles character-by-character output.
- Created a streaming server capable of processing requests with AI or mock responses.
- Added Docker support for both client and server with a docker-compose setup.
- Enhanced task handling with proper event streaming and error management.
- Updated tests to ensure context history handling and agent configuration are preserved.
