+++ examples/ai-powered/client/main.go
@@ -1,0 +1,80 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"log"
+	"os"
+	"time"
+
+	"github.com/inference-gateway/adk/client"
+	"github.com/inference-gateway/adk/types"
+	"go.uber.org/zap"
+)
+
+func main() {
+	// Get server URL from environment or use default
+	serverURL := os.Getenv("SERVER_URL")
+	if serverURL == "" {
+		serverURL = "http://localhost:8080/a2a"
+	}
+
+	// Initialize logger
+	logger, err := zap.NewDevelopment()
+	if err != nil {
+		log.Fatalf("Failed to create logger: %v", err)
+	}
+	defer func() {
+		if err := logger.Sync(); err != nil {
+			log.Printf("Failed to sync logger: %v", err)
+		}
+	}()
+
+	// Create client
+	a2aClient := client.NewClientWithLogger(serverURL, logger)
+
+	// Create a context with timeout
+	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
+	defer cancel()
+
+	// Prepare different types of messages to test AI capabilities
+	prompts := []string{
+		"What is the capital of France?",
+		"Write a haiku about programming",
+		"Explain quantum computing in simple terms",
+	}
+
+	for i, prompt := range prompts {
+		fmt.Printf("\n--- Request %d ---\n", i+1)
+		fmt.Printf("Sending: %s\n", prompt)
+
+		// Create message with proper structure
+		message := types.Message{
+			Role: "user",
+			Parts: []types.Part{
+				map[string]any{
+					"kind": "text",
+					"text": prompt,
+				},
+			},
+		}
+
+		// Send the message
+		params := types.MessageSendParams{
+			Message: message,
+		}
+
+		response, err := a2aClient.SendTask(ctx, params)
+		if err != nil {
+			log.Printf("Failed to send message %d: %v", i+1, err)
+			continue
+		}
+
+		// Display the response
+		if response.Result != nil {
+			responseJSON, _ := json.MarshalIndent(response.Result, "", "  ")
+			fmt.Printf("Response:\n%s\n", string(responseJSON))
+		}
+	}
+}
