--- examples/client/go.mod
@@ -1,20 +1,0 @@
-module client-example
-
-go 1.24.3
-
-require (
-	github.com/google/uuid v1.6.0
-	github.com/inference-gateway/adk v0.0.0
-	github.com/sethvargo/go-envconfig v1.3.0
-	go.uber.org/zap v1.27.0
-)
-
-require (
-	github.com/cloudevents/sdk-go/v2 v2.15.2 // indirect
-	github.com/json-iterator/go v1.1.12 // indirect
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
-	github.com/modern-go/reflect2 v1.0.2 // indirect
-	go.uber.org/multierr v1.10.0 // indirect
-)
-
-replace github.com/inference-gateway/adk => ../../
