--- examples/client/main.go
@@ -1,11 +1,0 @@
-package main
-
-func main() {
-	// Placeholder for the main function
-	// Review the examples in cmd/* directory
-	// The general approach is:
-	// 1. Initialize logger
-	// 2. Create configuration
-	// 3. Create a client
-	// 4. Call the A2A server
-}
