+++ examples/minimal/client/main.go
@@ -1,0 +1,77 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"log"
+	"os"
+	"time"
+
+	"github.com/inference-gateway/adk/client"
+	"github.com/inference-gateway/adk/types"
+	"go.uber.org/zap"
+)
+
+func main() {
+	// Get server URL from environment or use default
+	serverURL := os.Getenv("SERVER_URL")
+	if serverURL == "" {
+		serverURL = "http://localhost:8080/a2a"
+	}
+
+	// Initialize logger
+	logger, err := zap.NewDevelopment()
+	if err != nil {
+		log.Fatalf("Failed to create logger: %v", err)
+	}
+	defer func() {
+		if err := logger.Sync(); err != nil {
+			log.Printf("Failed to sync logger: %v", err)
+		}
+	}()
+
+	// Create client
+	a2aClient := client.NewClientWithLogger(serverURL, logger)
+
+	// Create a simple task
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+
+	// Create the message
+	message := types.Message{
+		Role: "user",
+		Parts: []types.Part{
+			map[string]any{
+				"kind": "text",
+				"text": "Hello, this is a test message. Please respond with a greeting.",
+			},
+		},
+	}
+
+	fmt.Println("Sending message to server...")
+
+	// Send the message using SendTask
+	params := types.MessageSendParams{
+		Message: message,
+	}
+
+	response, err := a2aClient.SendTask(ctx, params)
+	if err != nil {
+		log.Fatalf("Failed to send message: %v", err)
+	}
+
+	// Extract task from response
+	if response.Result == nil {
+		log.Fatal("No result in response")
+	}
+
+	// Pretty print the entire response
+	responseJSON, err := json.MarshalIndent(response.Result, "", "  ")
+	if err != nil {
+		log.Fatalf("Failed to marshal response: %v", err)
+	}
+
+	fmt.Println("\nReceived response:")
+	fmt.Println(string(responseJSON))
+}
