--- examples/server/.env.gateway.example
@@ -1,83 +1,0 @@
-
-# General settings
-ENVIRONMENT=production
-ALLOWED_MODELS=
-# Telemetry
-TELEMETRY_ENABLE=false
-TELEMETRY_METRICS_PORT=9464
-# Model Context Protocol (MCP)
-MCP_ENABLE=false
-MCP_EXPOSE=false
-MCP_SERVERS=
-MCP_CLIENT_TIMEOUT=5s
-MCP_DIAL_TIMEOUT=3s
-MCP_TLS_HANDSHAKE_TIMEOUT=3s
-MCP_RESPONSE_HEADER_TIMEOUT=3s
-MCP_EXPECT_CONTINUE_TIMEOUT=1s
-MCP_REQUEST_TIMEOUT=5s
-MCP_MAX_RETRIES=3
-MCP_RETRY_INTERVAL=5s
-MCP_INITIAL_BACKOFF=1s
-MCP_ENABLE_RECONNECT=true
-MCP_RECONNECT_INTERVAL=30s
-MCP_POLLING_ENABLE=true
-MCP_POLLING_INTERVAL=30s
-MCP_POLLING_TIMEOUT=5s
-MCP_DISABLE_HEALTHCHECK_LOGS=true
-# Agent-to-Agent (A2A) Protocol
-A2A_ENABLE=false
-A2A_EXPOSE=false
-A2A_AGENTS=
-A2A_CLIENT_TIMEOUT=30s
-A2A_POLLING_ENABLE=true
-A2A_POLLING_INTERVAL=1s
-A2A_POLLING_TIMEOUT=30s
-A2A_MAX_POLL_ATTEMPTS=30
-A2A_MAX_RETRIES=3
-A2A_RETRY_INTERVAL=5s
-A2A_INITIAL_BACKOFF=1s
-A2A_ENABLE_RECONNECT=true
-A2A_RECONNECT_INTERVAL=30s
-A2A_DISABLE_HEALTHCHECK_LOGS=true
-A2A_SERVICE_DISCOVERY_ENABLE=false
-A2A_SERVICE_DISCOVERY_NAMESPACE=
-A2A_SERVICE_DISCOVERY_POLLING_INTERVAL=30s
-# Authentication
-AUTH_ENABLE=false
-AUTH_OIDC_ISSUER=http://keycloak:8080/realms/inference-gateway-realm
-AUTH_OIDC_CLIENT_ID=inference-gateway-client
-AUTH_OIDC_CLIENT_SECRET=
-# Server settings
-SERVER_HOST=0.0.0.0
-SERVER_PORT=8080
-SERVER_READ_TIMEOUT=30s
-SERVER_WRITE_TIMEOUT=30s
-SERVER_IDLE_TIMEOUT=120s
-SERVER_TLS_CERT_PATH=
-SERVER_TLS_KEY_PATH=
-# Client settings
-CLIENT_TIMEOUT=30s
-CLIENT_MAX_IDLE_CONNS=20
-CLIENT_MAX_IDLE_CONNS_PER_HOST=20
-CLIENT_IDLE_CONN_TIMEOUT=30s
-CLIENT_TLS_MIN_VERSION=TLS12
-CLIENT_DISABLE_COMPRESSION=true
-CLIENT_RESPONSE_HEADER_TIMEOUT=10s
-CLIENT_EXPECT_CONTINUE_TIMEOUT=1s
-# Providers
-ANTHROPIC_API_URL=https://api.anthropic.com/v1
-ANTHROPIC_API_KEY=
-CLOUDFLARE_API_URL=https://api.cloudflare.com/client/v4/accounts/{ACCOUNT_ID}/ai
-CLOUDFLARE_API_KEY=
-COHERE_API_URL=https://api.cohere.ai
-COHERE_API_KEY=
-GROQ_API_URL=https://api.groq.com/openai/v1
-GROQ_API_KEY=
-OLLAMA_API_URL=http://ollama:8080/v1
-OLLAMA_API_KEY=
-OPENAI_API_URL=https://api.openai.com/v1
-OPENAI_API_KEY=
-DEEPSEEK_API_URL=https://api.deepseek.com
-DEEPSEEK_API_KEY=
-GOOGLE_API_URL=https://generativelanguage.googleapis.com/v1beta/openai
-GOOGLE_API_KEY=
