--- examples/server/.well-known/agent-card.json
@@ -1,88 +1,0 @@
-{
-  "name": "Weather Assistant",
-  "description": "A specialized AI agent that provides comprehensive weather information and forecasts",
-  "version": "2.1.0",
-  "url": "https://weather-agent.example.com",
-  "documentationUrl": "https://weather-agent.example.com/docs",
-  "iconUrl": "https://weather-agent.example.com/icon.png",
-  "capabilities": {
-    "streaming": true,
-    "pushNotifications": true,
-    "stateTransitionHistory": false
-  },
-  "defaultInputModes": ["text"],
-  "defaultOutputModes": ["text", "json"],
-  "skills": [
-    {
-      "id": "current-weather",
-      "name": "Current Weather",
-      "description": "Get current weather conditions for any location",
-      "tags": ["weather", "current", "conditions"],
-      "inputModes": ["text"],
-      "outputModes": ["text", "json"],
-      "examples": [
-        "What's the weather in New York?",
-        "Current conditions in Tokyo",
-        "Is it raining in London right now?"
-      ]
-    },
-    {
-      "id": "weather-forecast",
-      "name": "Weather Forecast",
-      "description": "Provide detailed weather forecasts for multiple days",
-      "tags": ["weather", "forecast", "prediction"],
-      "inputModes": ["text"],
-      "outputModes": ["text", "json"],
-      "examples": [
-        "5-day forecast for Paris",
-        "What will the weather be like this weekend in San Francisco?",
-        "Weekly weather outlook for Berlin"
-      ]
-    },
-    {
-      "id": "weather-alerts",
-      "name": "Weather Alerts",
-      "description": "Check for severe weather warnings and alerts",
-      "tags": ["weather", "alerts", "warnings", "severe"],
-      "inputModes": ["text"],
-      "outputModes": ["text", "json"],
-      "examples": [
-        "Any weather alerts for Miami?",
-        "Severe weather warnings in Texas",
-        "Hurricane updates for the Gulf Coast"
-      ]
-    }
-  ],
-  "provider": {
-    "organization": "Weather Corp",
-    "url": "https://weathercorp.example.com"
-  },
-  "securitySchemes": {
-    "apiKey": {
-      "type": "apiKey",
-      "in": "header",
-      "name": "X-API-Key",
-      "description": "API key for weather service access"
-    },
-    "bearer": {
-      "type": "http",
-      "scheme": "bearer",
-      "description": "Bearer token authentication"
-    }
-  },
-  "security": [
-    {
-      "apiKey": []
-    },
-    {
-      "bearer": []
-    }
-  ],
-  "additionalInterfaces": [
-    {
-      "transport": "websocket",
-      "url": "wss://weather-agent.example.com/ws"
-    }
-  ],
-  "supportsAuthenticatedExtendedCard": true
-}
