--- examples/server/main.go
@@ -1,12 +1,0 @@
-package main
-
-func main() {
-	// Placeholder for the main function
-	// Review the examples in cmd/* directory
-	// The general approach is:
-	// 1. Initialize logger
-	// 2. Create configuration
-	// 3. Create a toolbox with tools
-	// 4. Create server with configuration and toolbox
-	// 5. Start the server
-}
